Spring implantation to the project

+++ dao/src/main/java/com/gmail/sshekh/dao/DiscountDao.java
@@ -2,12 +2,13 @@ package com.gmail.sshekh.dao;
 
 import com.gmail.sshekh.dao.model.Discount;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface DiscountDao extends GenericDao<Discount> {
-    List<Discount> getDiscountsByRate(Integer discount);
+    List<Discount> getDiscountsByRate(BigDecimal discount);
 
-    Long getDiscountId(Integer discount);
+    Long getDiscountId(BigDecimal discount);
 
     Discount getDiscountById(Long id);
